Reorganize supporting types, add convenience inits to GazeButton
